Add files via upload
This is a code for a faulty calculator. In this faulty calculator, the user will get correct answers of simple arithmetic calculations, but they will get wrong answers for some specific questions which are selected by the examiner. This calculator is programmed in such a way that the exam questions, when putting in the calculator, will always give wrong answers.