Add files via upload
This is a small game, where the player has to guess a number between 1 and 25. He will get only 9 chances. Every time the players make a guess,  the game will tell whether he guessed correctly or not. Also, he/she gets a hint that the guess number is smaller or larger than the correct number. When all the guesses are used, GAME OVER!