Add files via upload
This is a program which prints a stair case of height between 1 to 8. It asks the user for height, then asks for true or false. If user gives true as answer, program will print upright staircase of input height. If the user gives false as answer, program print upside-down staircase of given height.